Tests working with caveats
